adding isDelinquent status for a loan

+++ db/queries/loans.sql
@@ -11,7 +11,7 @@ INSERT INTO payments (loan_id, week_number, amount, paid_at)
 VALUES ($1, $2, $3, $4)
 RETURNING *;
 -- name: GetLastPaidWeek :one
-SELECT COALESCE(MAX(week_number), 0)::BIGINT AS last_paid_week
+SELECT COALESCE(MAX(week_number), 0)::INT AS last_paid_week
 FROM payments
 WHERE loan_id = $1;
 -- name: InsertLoan :one
